refactor: Group related importation components in submodules
